Do not confuse "empty inventory" with "unknown inventory"
The fact that 0 may mean both "empty inventory" and "unknown inventory"
will confuse cache refreshing. If a trackback is removed from a cache,
refreshing it will show an empty inventory but a count of 1, because
the new count would be ignored as if it was undefined.

Until we find a better solution to handle inventory in one place
instead of two, copy inventory items count from old cache only
when we copy inventory content.

That should help with issue #344.